Refactor add and install commands to only require app name
The add and install commands previously required both app name and path from the user, even though the path was already stored in the config file. After the refactor, these commands now only require the app name. If a path is associated with the app name in the config, the command will retrieve and use it automatically